refactor(main-api): CommandQueryController 분리 및 패키지 이동
