+++ modules/infrastructure/api-query-jpa/src/main/java/com/whoz_in/api_query_jpa/member/MemberJpaViewer.java
@@ -1,0 +1,13 @@
+package com.whoz_in.api_query_jpa.member;
+
+import com.whoz_in.main_api.query.member.application.MemberAuthInfo;
+import com.whoz_in.main_api.query.member.application.MemberViewer;
+import java.util.Optional;
+
+public class MemberJpaViewer implements MemberViewer {
+
+    @Override
+    public Optional<MemberAuthInfo> getAuthInfoByLoginId(String loginId) {
+        return Optional.empty();
+    }
+}
