+++ modules/main-api/src/main/java/com/whoz_in/main_api/query/device_info/presentation/DeviceQueryController.java
@@ -3,18 +3,18 @@ package com.whoz_in.main_api.query.device_info.presentation;
 import com.whoz_in.main_api.query.device_info.application.DeviceInfoGet;
 import com.whoz_in.main_api.command.shared.application.CommandBus;
 import com.whoz_in.main_api.query.shared.application.QueryBus;
-import com.whoz_in.main_api.shared.presentation.CommandQueryController;
-import com.whoz_in.main_api.shared.presentation.response.CrudResponseCode;
-import com.whoz_in.main_api.shared.presentation.response.ResponseEntityGenerator;
+import com.whoz_in.main_api.query.shared.presentation.QueryController;
+import com.whoz_in.main_api.shared.presentation.CrudResponseCode;
+import com.whoz_in.main_api.shared.presentation.ResponseEntityGenerator;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-public class DeviceQueryController extends CommandQueryController {
+public class DeviceQueryController extends QueryController {
 
-    protected DeviceQueryController(CommandBus commandBus, QueryBus queryBus) {
-        super(commandBus, queryBus);
+    public DeviceQueryController(QueryBus queryBus) {
+        super(queryBus);
     }
 
     @GetMapping("/device")
