+++ modules/main-api/src/main/java/com/whoz_in/main_api/query/shared/presentation/QueryController.java
@@ -1,4 +1,4 @@
-package com.whoz_in.main_api.shared.presentation;
+package com.whoz_in.main_api.query.shared.presentation;
 
 import com.whoz_in.main_api.command.shared.application.Command;
 import com.whoz_in.main_api.command.shared.application.CommandBus;
@@ -9,14 +9,8 @@ import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor(access = AccessLevel.PROTECTED)
-public abstract class CommandQueryController {
-    private final CommandBus commandBus;
+public abstract class QueryController {
     private final QueryBus queryBus;
-
-    protected void dispatch(Command command){
-        commandBus.dispatch(command);
-    }
-
     /**
      * <p><b>주의</b></p>
      * Response는 추상 클래스가 아닌 인터페이스이기 때문에 아래와 같이 사용해도 경고만 발생하고 컴파일 에러는 발생하지 않습니다. </p>
