+++ modules/main-api/src/main/java/com/whoz_in/main_api/shared/presentation/ApiResponseBody.java
@@ -1,0 +1,5 @@
+package com.whoz_in.main_api.shared.presentation;
+
+public abstract class ApiResponseBody {
+
+}
