+++ modules/main-api/src/main/java/com/whoz_in/main_api/shared/presentation/ApiResponseCode.java
@@ -1,4 +1,4 @@
-package com.whoz_in.main_api.shared.presentation.response;
+package com.whoz_in.main_api.shared.presentation;
 
 import org.springframework.http.HttpStatus;
