--- modules/main-api/src/main/java/com/whoz_in/main_api/shared/presentation/response/ApiResponseBody.java
@@ -1,5 +1,0 @@
-package com.whoz_in.main_api.shared.presentation.response;
-
-public abstract class ApiResponseBody {
-
-}
